refactor: adapts all dao and services for using an Integer instead of Long, where id is needed

+++ src/main/java/com/pokemon/ao/persistence/dao/MoveDAO.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MoveDAO extends JpaRepository<Move, Long> {
+public interface MoveDAO extends JpaRepository<Move, Integer> {
 }
