+++ src/main/java/com/pokemon/ao/persistence/dao/PokemonDAO.java
@@ -1,7 +1,6 @@
 package com.pokemon.ao.persistence.dao;
 
 import com.pokemon.ao.persistence.entity.Pokemon;
-import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
@@ -9,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface PokemonDAO extends JpaRepository<Pokemon, Long> {
+public interface PokemonDAO extends JpaRepository<Pokemon, Integer> {
     @Query("SELECT p.id FROM Pokemon p")
-    List<Long> findAllIds();
+    List<Integer> findAllIds();
 }
