+++ src/main/java/com/pokemon/ao/persistence/service/PokemonService.java
@@ -11,13 +11,13 @@ import org.springframework.stereotype.Service;
 import java.util.List;
 
 @Service
-public class PokemonService extends AbstractService<PokemonVO, Pokemon, Long> {
+public class PokemonService extends AbstractService<PokemonVO, Pokemon, Integer> {
     @Autowired
-    protected PokemonService(Marshaller<PokemonVO, Pokemon> marshaller, JpaRepository<Pokemon, Long> dao) {
+    protected PokemonService(Marshaller<PokemonVO, Pokemon> marshaller, JpaRepository<Pokemon, Integer> dao) {
         super(marshaller, dao);
     }
 
-    public List<Long> findAllIds(){
+    public List<Integer> findAllIds(){
         PokemonDAO pokemonDAO = (PokemonDAO) dao;
         return pokemonDAO.findAllIds();
     }
